split to tky2jgd
